requirements.txt
fastapi
sqlalchemy
python-dotenv
uvicorn
psycopg2-binary